requirements.txt
aiohttp==3.8.4
aiopg==1.4.0
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
grequests==0.6.0
h11==0.14.0
idna==3.4
multidict==6.0.4
peewee==3.16.1
peewee-async==0.8.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
secure-json==1.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
